==> 1- preprocessing_genreAV_cluster.ipynb <==
import pandas as pd
import numpy as np
from sklearn.cluster import KMeans
import seaborn as sns
import sys
import sklearn.metrics as metrics
from AFSA import AFSA

np.set_printoptions(threshold=sys.maxsize)
import warnings
warnings.filterwarnings("ignore")

#Load the Ratings data
ratings_df = pd.read_csv('data/u.data', sep="\t", header=None)
ratings_df.columns = ['user_id', 'movie_id', 'rating', 'timestamp']
ratings = ratings_df.drop(['timestamp'], axis=1)
ratings

'''sort rating df by user id and item id'''
rating_sort = ratings.sort_values(by=['user_id', 'movie_id']).reset_index(drop=True)
rating_sort

#Load the Movies data
item = pd.read_csv('data/u.item', sep="|", encoding='latin-1', header=None)
item.columns = ['movie_id', 'movie title' ,'release date','video release date', 'IMDb URL', 'unknown', 'Action', 
                'Adventure', 'Animation', 'Children\'s', 'Comedy', 'Crime', 'Documentary', 'Drama', 'Fantasy', 'Film-Noir', 
                'Horror', 'Musical', 'Mystery', 'Romance', 'Sci-Fi', 'Thriller', 'War', 'Western']
item.head()

#drop unnecessary columns
item_genres01 = item.drop(['movie title', 'release date', 'video release date', 'IMDb URL', 'unknown' ], axis=1)
item_genres01

#export the output
item_genres01.to_csv('item_genres01.csv')

#merge ratings with items dataframe
user_genre = pd.merge(rating_sort,item_genres01)
user_genre

#replace the value 1 in each genres columns with the corresponding value from the rating column of the user_genre DataFrame

ratings1 = user_genre.replace({"Action":{1:user_genre.rating},"Adventure":{1:user_genre.rating},
                            "Animation":{1:user_genre.rating},"Children's":{1:user_genre.rating},"Comedy":{1:user_genre.rating},
                            "Crime":{1:user_genre.rating},"Documentary":{1:user_genre.rating},"Drama":{1:user_genre.rating},
                            "Fantasy":{1:user_genre.rating},"Film-Noir":{1:user_genre.rating},"Horror":{1:user_genre.rating},
                            "Musical":{1:user_genre.rating},"Mystery":{1:user_genre.rating},"Romance":{1:user_genre.rating},
                            "Sci-Fi":{1:user_genre.rating},"Thriller":{1:user_genre.rating},"War":{1:user_genre.rating},  
                            "Western":{1:user_genre.rating}})

#sort df by userID

ratings2 = ratings1.sort_values('user_id').drop(['movie_id', 'rating'], axis=1).reset_index(drop=True)  
ratings2

#Replace 0 values with NaN to avoid errors in calculating the average

ratings3 = ratings2.replace({"Action":{0:np.nan},"Adventure":{0:np.nan},
                            "Animation":{0:np.nan},"Children's":{0:np.nan},"Comedy":{0:np.nan},
                            "Crime":{0:np.nan},"Documentary":{0:np.nan},"Drama":{0:np.nan},
                            "Fantasy":{0:np.nan},"Film-Noir":{0:np.nan},"Horror":{0:np.nan},
                            "Musical":{0:np.nan},"Mystery":{0:np.nan},"Romance":{0:np.nan},
                            "Sci-Fi":{0:np.nan},"Thriller":{0:np.nan},"War":{0:np.nan},  
                            "Western":{0:np.nan}})
ratings3

#Get the average of users ratings

ratings4 = ratings3.groupby('user_id').mean().fillna(0)
ratings4

#export the output
ratings4.to_csv('AV_user_genre.csv')
