==> 3 - add_cluster_to_rates.ipynb <==
# Import packages
import pandas as pd

#import data

df_cluster1 = pd.read_csv('rating_genre_5cluster.csv')
df_cluster = df_cluster1[['user_id', 'cluster']]
df_cluster

#Load the Ratings data
ratings_df = pd.read_csv('data/u.data', sep="\t", header=None)
ratings_df.columns = ['user_id', 'movie_id', 'rating', 'timestamp']
ratings = ratings_df.drop(['timestamp'], axis=1)
ratings

# merge ratings with clusters data

user_cluster_rate = pd.merge(ratings,df_cluster).sort_values('user_id').set_index('user_id')
user_cluster_rate

# export the output data

user_cluster_rate.to_csv('user_5cluster_rate.csv')
